requirements.txt
flask
sqlmodel
strawberry-graphql